requirements.txt
kafka-python==1.3.3
oauthlib==2.0.2
pkg-resources==0.0.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
config==0.3.9
elasticsearch==5.3.0
kafka-python==1.3.3
nltk==3.2.2
oauthlib==2.0.2
pkg-resources==0.0.0
pymongo==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
textblob==0.12.0
tweepy==3.5.0
urllib3==1.20
config==0.3.9
elasticsearch==5.3.0
kafka-python==1.3.3
nltk==3.2.2
numpy==1.12.1
oauthlib==2.0.2
pkg-resources==0.0.0
pymongo==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
textblob==0.12.0
tweepy==3.5.0
urllib3==1.20
